CategoryCreationScreen architecture have been changed from MVP to MVVM
